third_party/blink/web_tests/svg/dynamic-updates/SVGFEPointLightElement-svgdom-y-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' property of the SVGFEPointLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pointLight.y.baseVal is 120
PASS pointLight.y.baseVal is 180
PASS successfullyParsed is true

TEST COMPLETE
